requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
aiofiles 